+++ package.json
@@ -24,17 +24,16 @@
   },
   "homepage": "https://github.com/Nicolas-Pistillo/graph_api#readme",
   "dependencies": {
+    "apollo-server-express": "^3.3.0",
     "colors": "^1.4.0",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "express-graphql": "^0.12.0",
     "graphql": "^15.5.2",
     "graphql-import-node": "0.0.4",
     "graphql-playground-middleware-express": "^1.7.22",
     "graphql-tools": "^8.1.0",
     "http": "0.0.1-security",
-    "ncp": "^2.0.0",
     "typescript": "^4.4.2"
   },
   "devDependencies": {
