+++ src/resolvers/resolversMap.ts
@@ -1,0 +1,10 @@
+import { IResolvers } from "@graphql-tools/utils";
+import query from "./query";
+
+const resolvers : IResolvers = {
+
+    ...query
+
+}
+
+export default resolvers;
