+++ src/schema/schema.graphql
@@ -1,0 +1,6 @@
+type Query {
+    hola: String!
+    holaConNombre(nombre: String!): String!
+    holaConGraphQl: String!
+    testNick: String!
+}
